pyproject.toml
[project]
name = "fake-llm-server"
version = "0.1.0"
description = "Fake (test double) LLM server"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "huggingface-hub>=1.3.5",
    "llama-cpp-python>=0.3.16",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "openai>=2.16.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "types-requests>=2.32.4.20260107",
    "types-setuptools>=80.10.0.20260124",
]

[build-system]
requires = ["uv_build>=0.10"]
build-backend = "uv_build"

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/__init__.py",
]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-config=pyproject.toml -ra --quiet"
pythonpath = "src"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812"]  # Conflicts with the formatter.

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # assert is idiomatic in Pytest.

[tool.ruff.lint.pydocstyle]
convention = "google"
